requirements.txt
gradio>=4.0.0
langchain>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0
accelerate
pypdf2
python-dotenv
numpy
pandas
tiktoken
faiss-cpu
huggingface-hub
requests
urllib3